+++ __tests__/controllers/routers/driver.test.js
@@ -1,4 +1,5 @@
 // module
+const { _fn } = require('../../../controllers/helpers');
 const { asyncWrapper, getMethods,
     getMiddlewareQueue, getControllerQueue, getViewRenderQueue } = require('../../../controllers/routers/driver')._test;
 
@@ -12,24 +13,24 @@ describe('Check the ENV', () => {
 });
 
 
-// describe('Fn: asyncWrapper', () => {
-//     test('Should wrap asyncfunctions with an error catcher', () => {
-//         const mockInput = [async (req, res, next) => {}, (req, res, next) => {}];
-//         const result = mockInput.map(fn => asyncWrapper(fn));
-//         //
-//         expect(checkNativeBrand(result[0], 'array')).toBeTruthy();
-//         expect(checkNativeBrand(result[1], 'array')).toBeTruthy();
-//         expect(result[0].toString()).toBe('(req, res, next) => fn(req, res, next).catch(next)');
-//         expect(result[1].toString()).toBe('(req, res, next) => {}');
-//     });
-// });
+describe('Fn: asyncWrapper', () => {
+    test('Should wrap asyncfunctions with an error catcher', () => {
+        const mockInput = [async (req, res, next) => {}, (req, res, next) => {}];
+        const result = mockInput.map(fn => asyncWrapper(fn));
+        //
+        expect(_fn.object.checkNativeBrand(result[0], 'array')).toBeTruthy();
+        expect(_fn.object.checkNativeBrand(result[1], 'array')).toBeTruthy();
+        expect(result[0].toString()).toBe('(req, res, next) => fn(req, res, next).catch(next)');
+        expect(result[1].toString()).toBe('(req, res, next) => {}');
+    });
+});
 
 
 describe('Bundle: router.js', () => {
     const mockMethod = ['get', 'pull'];
     const mockController1 = { get: () => { return false }, pull: () => { return true } };
     const mockController2 = [() => { return 0 }, () => { return 1 }];
-    const mockSetting1 = { title: 'test', authentication: true, authorization: true, template: './test' };
+    const mockSetting1 = { title: 'test', authenticated: true, authorized: true, template: './test' };
     const mockSetting2 = { title: 'test', titleOption: { root: true }, crawler: false };
 
     test('Fn: getMethods: Should normalize into a method array in the anti-alphabetical order', () => {
@@ -44,8 +45,8 @@ describe('Bundle: router.js', () => {
         const test = [mockSetting1, mockSetting2];
         const result = test.map(obj => getMiddlewareQueue(obj));
         //
-        expect(result[0].length).toBe(6);
-        expect(result[1].length).toBe(2);
+        expect(result[0].length).toBe(7);
+        expect(result[1].length).toBe(3);
     });
 
     test('Fn: getControllerQueue: Should stack a queue from controllers with normalization', () => {